Multiple colours for bubbles
For now: cyan, red, and green.

This sorts out a mechanism for laying out the tile/panel map so that
each of the colour combinations is covered.

Bubbles are placed into a 4x4 grid:
 - Along the Y axis is the bubble to draw {none, cyan, red, green}
 - Along the X axis is the neighbour colour {none, cyan, red, green}

The appropriate combinations are selected when painting tile data into
the strips that make up the game board.