Mapping pixel coordinates to game-board position
For now this is placed in the debug-cursor code, but will eventually be
used as the first step of bubble collision detection.

The mapping follows two steps:

 1. Imagine that the grid is rectangles, like a brick wall. This makes it
    easy to calculate a starting-point for the game-board coordinate as
    it's just offsets and division.

 2. From the pixel position within the rectangle, index an array to
    get the correction value, which maps the pixel to its proper hex-grid
    position.

In addition, the set_bubble function now takes in a game-board position
rather than an (x, y) position. This does lead to a bit of doing and then
undoing math, so could be refactored again in the future if it becomes
a problem.