Collision map for fast collision detection
Similar to the array pixel_to_board, which maps a pixel within a 16x14
staggered-grid into a relative game-board position:

This new array contains bit-fields for each of the six neighbours, set
if a bubble centred on the pixel collides with the neighbour.

'centre' is a little ambiguous, as the sprite is an even number of pixels
in both directions, so there are kinda four centre pixels. For the purpose
of this array, the chosen centre pixel is (7, 7)