Bubble collisions
Bubbles can now collide and stick to one another.

For now, the debug-cursor code is removed, although the sprite remains.

Collisions aren't perfect, in at least one case a bubble was seen passing
through one bubble on the game-board and colliding with the next one,
then calling set_bubble for the position of the bubble it first passed
though. Therefore, any number of bubbles launched at that particular
angle would never result in additional bubbles added to the game board.

The bubbles also, following what Bust-A-Move 2 does on the PS1, are
allowed to pass a few pixels into the bubble they collide with, for a
frame.

This seems to look a bit choppy on the SMS. Maybe it will look better
once some extra frames smooth the transition from the overlap into the
final game-board position. If not, then possibly the frame of overlap
should be removed.