Bubble popping detection
When a bubble reaches the end of its flight, a new function checks if the
bubble has formed a group of three or more.

The check is done using a stack to implement a depth-first search.
A map of the game board, match_map, keeps track of which game-board
positions have already been processed and which have already been added
to the stack.