Bubble-drop detection
Not animated yet, the bubbles just disappear. But detection of which
bubbles should drop is now implemented using the same mechanism as when
detecting groups of 3+ bubbles.

Bubbles in the first row are added to the search stack. Instead of looking
for same-colour bubbles, the search looks only for present bubbles. After
completing the search, any that have not reached the FLOAT_CONFIRMED state
are assumed to be floating without support so should be dropped.