Use randomised bubbles, of colours currently on the game board
Keep track of how many of each colour are currently on the game-board.
When loading the launcher, generate random colours until a colour comes
up that is present on the game board.

Not quite right yet however: Bubbles waiting in the fall-queue are also
still on the game-board, so an extinct bubble can be loaded if it was
about to fall.