Separate game_board and game_board_visible for logical vs display use
game_board is updated as soon as a bubble is in-play / out-of-play.
game_board_visible represents the bitmap, and is only updated when a
bubble is added / removed from the bitmap.

This solves the problem where generating the next random bubble,  it
could pick an extinct colour that was waiting for its fall animation
to begin.