Bit-field to mark bubbles as invalid
Previously, bubbles were being marked invalid in wash_bubbles_grey by
setting them to BUBBLE_MAX. This caused corruption when painting the
next level due to the neighbouring bubbles affecting which pattern is
used, and some neighbours not being populated with valid values yet.

Using a specific bit means that it can be basked off when selecting
which pattern to draw.