Move PSGlib callbacks into frame interrupt handler
SMS_waitForVBlank simply waits for the next frame interrupt to go off
before returning. For this to give correct frame pacing requires that
all of one frame's processing is completed before the next interrupt
goes off.

This is not always the case however. If the work done for one frame of
gameplay ends up taking too long, then a VBlank interrupt is 'missed',
adding jitter (in the form of missed frames) to the rate we call into
PSGlib. This is very noticeable for sound effects.

Some parts of the code need to be improved to prevent skipped frames in
general. Eg, right now, selecting the next bubble colour can take a long
time if it keeps having to retry due to picking not-on-board colours.

For now however, the audio jitter can be fixed by moving the call into
an interrupt handler, which will be called every frame, even when the
gameplay loop ends up missing a frame.